requirements.txt
asgiref==3.5.2
boto3==1.26.29
botocore==1.29.29
cffi==1.15.1
cryptography==38.0.4
dj-database-url==1.2.0
Django==4.1.3
django-environ==0.9.0
django-heroku==0.3.1
django-rest-knox==4.2.0
django-storages==1.13.1
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
